Restructured files to match original layout. Still stuggling with tokens having invalid signatures.
