+++ knexfile.js
@@ -1,18 +1,14 @@
 // Update with your config settings.
 
 module.exports = {
-
   development: {
     client: 'sqlite3',
-    connection: {
-      filename: './data/users.sqlite3'
-    },
-    useNullAsDefault: true,
+    connection: { filename: './database/auth.db3' }, // change this if you want a different name for the database
+    useNullAsDefault: true, // used to avoid warning on console
     migrations: {
-      directory: './data/migrations'
+      directory: './database/migrations',
+      tableName: 'dbmigrations',
     },
-    seeds: {
-      directory: './data/seeds'
-    }
-  }
+    seeds: { directory: './database/seeds' },
+  },
 };
