+++ package.json
@@ -14,6 +14,7 @@
     "cors": "^2.8.5",
     "debug": "^4.1.1",
     "express": "^4.16.4",
+    "helmet": "^3.15.0",
     "jsonwebtoken": "^8.4.0",
     "knex": "^0.16.3",
     "sqlite3": "^4.0.6"
